requirements.txt
asgiref==3.8.1
boto3==1.34.138
botocore==1.34.138
contourpy==1.3.1
cycler==0.12.1
Django==5.0.6
django-cors-headers==4.3.1
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
et_xmlfile==2.0.0
fonttools==4.56.0
gunicorn==22.0.0
inflection==0.5.1
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
lxml==5.3.1
matplotlib==3.10.1
mysqlclient==2.2.4
numpy==2.2.4
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.1.0
psycopg2==2.9.10
PyJWT==1.7.1
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
s3transfer==0.10.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.16.0
sqlparse==0.5.0
threadpoolctl==3.6.0
typing_extensions==4.13.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wordcloud==1.9.4